test(style): adding TU (90% code cov now)
