+++ src/style/istyle-manager.js
