+++ test/style/sass.manager.spec.js
@@ -2,62 +2,36 @@ import { expect } from 'chai'
 import 'mocha'
 import { join } from 'path'
 
-import SassStyle from '../../src/style/compiler/sass.manager.js'
+import StyleManager from '../../src/style/manager.js'
+import SassStyle from '../../src/style/sass/manager.js'
 
-describe('Sass Style', () => {
-  const sassStyle = new SassStyle()
-  const testFiles = [
-    {
-      expectMatch: true,
-      expectResolver: join('/', 'css', 'main.css'),
-      file: 'main.scss'
-    },
-    {
-      expectMatch: false,
-      expectResolver: join('/', 'css', 'main.css'),
-      file: 'main.css'
-    },
-    {
-      expectMatch: false,
-      expectResolver: join('/', 'css', 'main.css'),
-      file: 'main.scsss'
-    },
-    {
-      expectMatch: false,
-      expectResolver: join('/', 'css', '.css'),
-      file: ''
-    }
-  ]
-  testFiles.forEach((test) => {
-    it('matchExtension - ' + test.file, () => {
-      expect(sassStyle.isMatchedExtension(test.file))
-        .to.equal(test.expectMatch)
-    })
-    it('resolver - ' + test.file, () => {
-      expect(sassStyle.resolver(test.file))
-        .to.equal(test.expectResolver)
-    })
+describe('style manager', () => {
+  // set input variables.
+  const themePath = join(__dirname, '..', 'fixtures')
+  const publicPath = join(__dirname, '..', 'fixtures')
+
+  const manager = new StyleManager(themePath, publicPath)
+
+  it('should test constructor', () => {
+    expect(manager.themePath).to.equal(themePath)
+    expect(manager.publicPath).to.equal(publicPath)
   })
 
-  const testsOutput = [
-    {
-      fileDirs: ['feature', 'index.scss'],
-      publicPath: 'home/username/public',
-      expect: {
-        fullFileName: 'index.scss',
-        dist: join('home', 'username', 'public', 'css', 'feature', 'index.css')
-      }
-    }
-  ]
-
-  testsOutput.forEach((test) => {
-    it('destinationCompiledFile - ' + test.fileDirs.toString(), () => {
-      const res = sassStyle.destinationCompiledFile(test.fileDirs, test.publicPath)
-      expect(res.fullFileName)
-        .to.equal(test.expect.fullFileName)
-      expect(res.dist)
-        .to.equal(test.expect.dist)
-    })
+  it('should check the compiler(s) set', () => {
+    manager.setCompilers()
+    expect(manager.compilers.length).to.equal(1)
+    expect(manager.compilers[0] instanceof SassStyle).to.equal(true)
   })
 
+  it('should check compile', () => {
+    const matchedCompilers = manager.compile('main.scss')
+    expect(matchedCompilers.length).to.equal(1)
+    expect(matchedCompilers[0] instanceof SassStyle).to.equal(true)
+  })
+
+  it('should resolve compiled file', () => {
+    const filename = 'main.scss'
+    const resolvedPath = manager.resolver(filename)
+    expect(resolvedPath).to.equal((new SassStyle()).resolver(filename))
+  })
 })
